검색결과-기업추천/비추천 코드 수정-투표 후 기업의 선호도 출력값을 추천클릭수에서 퍼센티지(%)로  변경(소수점이하 안나오도록 추가설정 필요)

+++ blind/src/main/java/com/ksinfo/blind/search/dto/companyVoteResultDto.java
@@ -2,21 +2,22 @@ package com.ksinfo.blind.search.dto;
 
 public class companyVoteResultDto {
 
-	private int voteCountOfGood;
-	private int voteCountOfBad;
+	private float voteCountOfGood;
+	private float voteCountOfBad;
 	
-	public int getVoteCountOfGood() {
+	public float getVoteCountOfGood() {
 		return voteCountOfGood;
 	}
-	public void setVoteCountOfGood(int voteCountOfGood) {
+	public void setVoteCountOfGood(float voteCountOfGood) {
 		this.voteCountOfGood = voteCountOfGood;
 	}
-	public int getVoteCountOfBad() {
+	public float getVoteCountOfBad() {
 		return voteCountOfBad;
 	}
-	public void setVoteCountOfBad(int voteCountOfBad) {
+	public void setVoteCountOfBad(float voteCountOfBad) {
 		this.voteCountOfBad = voteCountOfBad;
 	}
+	
 
 
 }
